feat(api): add support for instance variables
